requirements.txt
python-kasa
flask[async]